penyesuaian di bagian back-end

+++ delcafe/database/migrations/2025_04_22_074530_create_pemesanan_table.php
@@ -15,7 +15,7 @@
                     $table->id();
                     $table->unsignedBigInteger('id_pelanggan');
                     $table->unsignedBigInteger('admin_id')->nullable();
-                    $table->integer('total_harga');
+                    $table->decimal('total_harga', 10, 2);
                     $table->enum('metode_pembayaran', ['tunai', 'qris', 'transfer bank'])->default('tunai');
                     $table->string('bukti_pembayaran')->nullable();
                     $table->enum('status', ['menunggu', 'pembayaran', 'dibayar', 'diproses', 'selesai', 'dibatalkan'])->default('menunggu');
